big update
update/create done
remove api part from front user controller

+++ RubymineProjects/Login/app/controllers/events_controller.rb
@@ -1,14 +1,15 @@
 class EventsController < ApplicationController
 
+  require 'ErrorMessage'
+
   respond_to :json
 
-  skip_before_action :verify_authenticity_token #ta bort
+  #not in use for API part
+  skip_before_action :verify_authenticity_token
 
-  #check for api key
-  #before_action :api_key
-  before_action :api_authenticate, only: [:new, :create, :update]
-  #check if api key is valid
-  before_action :restrict_access, only: [:index, :show]
+  #JWT Json Web Tokens
+  before_action :api_authenticate, only: [:index, :show, :new, :create, :edit, :update]
+
 
   # Checking if user want own limit/offset - definied in application_controller
   # for wider reach
@@ -16,31 +17,9 @@ class EventsController < ApplicationController
 
 
   def index
-
-    # Check the parameters
-    if params[:long].present? && params[:lat].present?
-
-      # searching by long lat (near by position)
-      positions = Position.near([params[:lat].to_f, params[:long].to_f], 20).limit(@limit).offset(@offset)
-      @events = positions.flat_map(&:events)
-
-    # Check the parameters
-    elsif params[:search].present?
-
-      # searching by event name
-      @events = Event.search(params[:search]).limit(@limit).offset(@offset).order('created_at DESC')
-    # Check the parameters
-    elsif params[:tag_search].present?
-
-      # searching by tags name
-      tags = Tag.search(params[:tag_search]).limit(@limit).offset(@offset).order('created_at DESC')
-      @events = tags.flat_map(&:events)
-
-    else
       @events = Event.limit(@limit).offset(@offset).order('created_at DESC')
-    end
-
   end
+
 
   def new
     @event = Event.new
@@ -77,22 +56,51 @@ class EventsController < ApplicationController
     render json: { error: error }, status: :bad_request # just json in this example
   end
 
-  def destroy
+  def edit
     event = Event.find(params[:id])
-
-    #destroy = all child included
-    if event.destroy
+    if event.present?
       @event = event
     end
 
       # render or error message
   rescue ActiveRecord::RecordNotFound
-    error = ErrorMessage.new('Could not find any resources. Bad parameters?', 'Could not delete any event!' )
+    error = ErrorMessage.new('Could not find any resources. Bad parameters?', 'Could not edit any event!' )
     render json: { error: error }, status: :bad_request # just json in this example
 
   end
 
+  def update
+    event = Event.find(params[:id])
+    if event.present?
+      # Handle a successful update.
 
+      #update event: name and description
+      event.update_attributes(event_params.except(:tag, :position))
+
+      @event = event
+    end
+
+    # render or error message
+      rescue ActiveRecord::RecordNotFound
+      error = ErrorMessage.new('Could not find any resources. Bad parameters?', 'Could not update any event!' )
+      render json: { error: error }, status: :bad_request # just json in this example
+    end
+
+  def destroy
+    event = Event.find(params[:id])
+
+    #destroy = all child included
+    if event.present?
+      if event.destroy
+        @event = event
+      end
+    end
+
+                # render or error message
+  rescue ActiveRecord::RecordNotFound
+    error = ErrorMessage.new('Could not find any resources. Bad parameters?', 'Could not delete any event!' )
+    render json: { error: error }, status: :bad_request # just json in this example
+  end
 
   private
 
@@ -108,32 +116,6 @@ class EventsController < ApplicationController
     json_params.require(:event).permit(:creator_id, :name, :description,
                                        position:[:long, :lat],
                                        tag:[:name])
-    #params.require(:event).permit(:name, :description)
-
-  end
-
-
-end
-
-################################### Custom class
-# This is a custom class for handling errors - Should be in another file!
-# No support from rails base model
-class ErrorMessage
-
-  def initialize(dev_mess, usr_mess)
-    # This is going to be json...camelcase
-    @developerMessage = dev_mess
-    @userMessage = usr_mess
-  end
-
-  # This is a custom class so we dont have the xml serializer included.
-  # I wrote an own to_xml (will be called by framework)
-  # There is probably a gem for that!?!
-  def to_xml(options={})
-    str = "<error>"
-    str += "  <developerMessage>#{@developerMessage}</developerMessage>"
-    str += "  <userMessage>#{@userMessage}</userMessage>"
-    str += "</error>"
   end
 
 end
